Refs #33003 - Rename files to new folder structure
