+++ webpack/assets/javascripts/react_app/components/FactCharts/constants.js
