+++ webpack/assets/javascripts/react_app/components/FactCharts/selectors.js
