+++ webpack/assets/javascripts/react_app/components/FactCharts/style.scss
